chore(cleanup): complete catalog consolidation by moving plugin manifest and updating test imports
