+++ catalog/json-plugins/plugin-manifest.json
